requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Demand Forecasting Pipeline — Requirements
#
# Install all:      pip install -r requirements.txt
# GPU support:      pip install torch --index-url https://download.pytorch.org/whl/cu118
# Chronos only:     pip install chronos-forecasting torch
# PyTorch models:   pip install pytorch-forecasting lightning
# ─────────────────────────────────────────────────────────────────────────────

# Core data
numpy>=1.24
pandas>=2.0
PyYAML>=6.0
python-dateutil>=2.8

# Chronos (Amazon pre-trained model)
# Install separately if only running Chronos:
# pip install git+https://github.com/amazon-science/chronos-forecasting.git
chronos-forecasting>=1.3

# PyTorch (Chronos + pytorch-forecasting)
torch>=2.0
# For CUDA GPU: replace above with:
# pip install torch --index-url https://download.pytorch.org/whl/cu118

# PyTorch Forecasting (TFT, NHiTS, DeepAR)
pytorch-forecasting>=1.0
lightning>=2.0

# Experiment tracking
mlflow>=2.10

# Optional: local secrets management
python-dotenv>=1.0

# Optional: Databricks (only needed on cluster, pre-installed)
# pyspark>=3.4
